requirements.txt
flask
requests
faiss-cpu
numpy
sentence-transformers